feat(hook): make mcp and skills hook based
