+++ app/daemon/src/gateway/builder.rs
@@ -1,8 +1,6 @@
 //! Hook builder — constructs a fully-configured GatewayHook from DaemonConfig.
 
 use crate::config;
-use crate::feature::mcp::McpHandler;
-use crate::feature::skill::SkillHandler;
 use crate::gateway::GatewayHook;
 use anyhow::Result;
 use memory::{InMemory, Memory};
@@ -11,14 +9,15 @@ use runtime::{Runtime, Tool};
 use std::path::Path;
 use std::sync::Arc;
 
-/// Build a fully-configured `Runtime<GatewayHook>` from config and directory.
+/// Build a fully-configured `Runtime` from config and directory.
 ///
-/// Constructs GatewayHook with all backends (model, memory, skills, MCP),
-/// then wraps it in a Runtime with loaded agents.
+/// Constructs GatewayHook with all backends (memory, skills, MCP),
+/// creates the model provider separately, then wraps both in a Runtime
+/// with loaded agents.
 pub async fn build_runtime(
     config: &crate::DaemonConfig,
     config_dir: &Path,
-) -> Result<Runtime<GatewayHook>> {
+) -> Result<Runtime<ProviderManager, GatewayHook>> {
     // Construct in-memory backend.
     let memory = InMemory::new();
     tracing::info!("using in-memory backend");
@@ -32,19 +31,30 @@ pub async fn build_runtime(
 
     // Load skills.
     let skills_dir = config_dir.join(config::SKILLS_DIR);
-    let skills = SkillHandler::load(skills_dir)?;
+    let skills = skill::SkillHandler::load(skills_dir)?;
 
-    // Load MCP servers.
-    let mcp = McpHandler::load(config_dir.to_path_buf(), &config.mcp_servers).await;
+    // Load MCP servers — convert daemon config to mcp config.
+    let mcp_configs: Vec<mcp::McpServerConfig> = config
+        .mcp_servers
+        .iter()
+        .map(|s| mcp::McpServerConfig {
+            name: s.name.clone(),
+            command: s.command.clone(),
+            args: s.args.clone(),
+            env: s.env.clone(),
+            auto_restart: s.auto_restart,
+        })
+        .collect();
+    let mcp_handler = mcp::McpHandler::load(config_dir.to_path_buf(), &mcp_configs).await;
 
     // Build GatewayHook.
-    let mut hook = GatewayHook::new(manager, memory, skills, mcp);
+    let mut hook = GatewayHook::new(memory, skills, mcp_handler);
 
     // Register memory tools on the hook.
     register_memory_tools(&mut hook);
 
-    // Wrap in Runtime.
-    let runtime = Runtime::new(Arc::new(hook));
+    // Wrap in Runtime — model and hook are separate.
+    let runtime = Runtime::new(manager, Arc::new(hook));
 
     // Load agents from markdown files.
     let agents = crate::loader::load_agents_dir(&config_dir.join(config::AGENTS_DIR))?;
