+++ crates/mcp/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "walrus-mcp"
+description = "Walrus MCP bridge"
+documentation = "https://docs.rs/walrus-mcp"
+readme = "README.md"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+keywords.workspace = true
+
+[dependencies]
+runtime.workspace = true
+wcore.workspace = true
+
+# crates-io dependencies
+anyhow.workspace = true
+compact_str.workspace = true
+serde.workspace = true
+rmcp.workspace = true
+schemars.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+tracing.workspace = true
