+++ crates/skill/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "walrus-skill"
+description = "Walrus skill registry"
+documentation = "https://docs.rs/walrus-skill"
+readme = "README.md"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+keywords.workspace = true
+
+[dependencies]
+runtime.workspace = true
+wcore.workspace = true
+
+# crates-io dependencies
+anyhow.workspace = true
+compact_str.workspace = true
+serde.workspace = true
+serde_yaml.workspace = true
+tokio.workspace = true
+tracing.workspace = true
